refactor: chat controller
