+++ src/controllers/admin/index.js
@@ -5,3 +5,4 @@ export * from "./users.controllers.js";
 export * from "./business.controllers.js";
 export * from "./analytics.controllers.js";
 export * from "./superadmin.controllers.js";
+export * from "./chat.controllers.js";
