+++ src/controllers/chat/index.js
@@ -1,0 +1,5 @@
+export * from './chat.management.controllers.js';
+export * from './message.controllers.js';
+export * from './message.deletion.controllers.js';
+export * from './message.features.controllers.js';
+export * from './chat.settings.controllers.js';
